Fixed a couple simple bugs in Box.

+++ src/main/java/bits/math3d/geom/ExtrudedLineLoop.java
@@ -67,11 +67,11 @@ public class ExtrudedLineLoop implements Volume {
     }
 
 
-    public double minZ() {
+    public float minZ() {
         return mMinZ;
     }
 
-    public double maxZ() {
+    public float maxZ() {
         return mMaxZ;
     }
